Create README.md
In this particular project we have coded an Admin Dashboard all based on Html, css and  a little bit of javascript in order to add some interaction on the project. It consists on  ADM dashboard to manipulate data where we're able to observe the cash-flow of  an University pointing out the number of employees as well as the l all money that in and out along the year.
I hope you enjoy it.